Import feature for customer and package

+++ app/Http/Controllers/ImportController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Imports\MembersImport;
+use App\Imports\PackagesImport;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 use Maatwebsite\Excel\Facades\Excel;
@@ -22,6 +23,26 @@ class ImportController extends Controller
             $imported_file->move('import', $file_name);
 
             Excel::import(new MembersImport, public_path('import/' . $file_name));
+
+            Storage::delete(public_path('import/' . $file_name));
+
+            return redirect()->back()->with('success', 'Import successful!');
+        }
+    }
+
+    public function importPackages(Request $request)
+    {
+        $request->validate([
+            'file' => ['required', 'mimes:csv,xlsx,xls']
+        ]);
+
+        $imported_file = $request->file('file');
+
+        if ($imported_file != null) {
+            $file_name = $imported_file->getClientOriginalName();
+            $imported_file->move('import', $file_name);
+
+            Excel::import(new PackagesImport, public_path('import/' . $file_name));
 
             Storage::delete(public_path('import/' . $file_name));
